+++ src/products/models/index.ts
@@ -3,3 +3,4 @@ export { ProductAttributeOption } from './product-attribute-option';
 export { ProductAttributes } from './product-attributes';
 export { ExtensionAttributes } from './extension-attributes';
 export { StockItem } from './stock-item';
+export { ProductLink } from './product-link';