+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-### 0.6.1
+### 0.7.0
+
+-   Added helper method for orders.
+-   Added methods to search orders.
+-   Added ability to query for linked products for a product.
+
+#### 0.6.1
 
 -   Added missing `ExtensionAttributes` and helper methods for stock querying.